requirements.txt
streamlit==1.26.0
openai>=1.2.0
python-dotenv==1.0.0
PyPDF2==3.0.1